services/logistics-service/requirements.txt
flask>=3.1.3
flask-cors>=6.0.2
kafka-python>=2.3.0
psycopg2-binary>=2.9.6
requests>=2.31.0